feat: persistent styles
found a way to implement persistent styles without FOUC/FART! i manually create a CSS string, which is inserted in `head` with `set:html` directive. then a tiny inline script checks if there was a saved style.

instead, when applying a style, a transition is added (client-side).
very please with the result!
closes #2